+++ AAEmu.Game/Core/Packets/L2G/LGPlayerReconnectPacket.cs
@@ -13,7 +13,7 @@ namespace AAEmu.Game.Core.Packets.L2G
 
         public override void Read(PacketStream stream)
         {
-            var accountId = stream.ReadUInt32();
+            var accountId = stream.ReadUInt64();
             var connection = GameConnectionTable.Instance.GetConnection(accountId);
             connection?.SendPacket(new SCReconnectAuthPacket(connection.Id));
         }
