+++ AAEmu.Game/Models/Game/Char/Character.cs
@@ -63,7 +63,7 @@ namespace AAEmu.Game.Models.Game.Char
         public List<IDisposable> Subscribers { get; set; }
 
         public uint Id { get; set; }
-        public uint AccountId { get; set; }
+        public ulong AccountId { get; set; }
         public Race Race { get; set; }
         public Gender Gender { get; set; }
         public short LaborPower { get; set; }
@@ -1572,7 +1572,7 @@ namespace AAEmu.Game.Models.Game.Char
                 SendPacket(packet);
         }
 
-        public static Character Load(uint characterId, uint accountId)
+        public static Character Load(uint characterId, ulong accountId)
         {
             using (var connection = MySQL.CreateConnection())
                 return Load(connection, characterId, accountId);
@@ -1608,7 +1608,7 @@ namespace AAEmu.Game.Models.Game.Char
 
         #region Database
 
-        public static Character Load(MySqlConnection connection, uint characterId, uint accountId)
+        public static Character Load(MySqlConnection connection, uint characterId, ulong accountId)
         {
             Character character = null;
             using (var command = connection.CreateCommand())
@@ -1735,7 +1735,7 @@ namespace AAEmu.Game.Models.Game.Char
                         character = new Character(modelParams);
                         character.Position = new Point();
                         character.Id = reader.GetUInt32("id");
-                        character.AccountId = reader.GetUInt32("account_id");
+                        character.AccountId = reader.GetUInt64("account_id");
                         character.Name = reader.GetString("name");
                         character.AccessLevel = reader.GetInt32("access_level");
                         character.Race = (Race)reader.GetByte("race");
